+++ __tests__/services/connection.service.spec.ts
@@ -8,7 +8,8 @@ import {
   DataChunkDataType,
   InworldPacket as ProtoPacket,
 } from '../../proto/packets.pb';
-import { SessionToken } from '../../src/common/interfaces';
+import { SessionToken } from '../../src/common/data_structures';
+import { protoTimestamp } from '../../src/common/helpers';
 import {
   CHAT_HISTORY_TYPE,
   HistoryItem,
@@ -17,6 +18,7 @@ import {
 import { GrpcAudioPlayback } from '../../src/components/sound/grpc_audio.playback';
 import { GrpcWebRtcLoopbackBiDiSession } from '../../src/components/sound/grpc_web_rtc_loopback_bidi.session';
 import { WebSocketConnection } from '../../src/connection/web-socket.connection';
+import { InworldPacket } from '../../src/entities/inworld_packet.entity';
 import { EventFactory } from '../../src/factories/event';
 import { ConnectionService } from '../../src/services/connection.service';
 import { WorldEngineService } from '../../src/services/world_engine.service';
@@ -259,7 +261,7 @@ describe('open', () => {
       sessionId: v4(),
       token: v4(),
       type: 'Bearer',
-      expirationTime: new Date().toISOString(),
+      expirationTime: protoTimestamp(),
     };
 
     const generateSessionToken = jest.fn(() => Promise.resolve(expiredSession));
@@ -482,7 +484,7 @@ describe('send', () => {
     jest
       .spyOn(GrpcAudioPlayback.prototype, 'excludeCurrentInteractionPackets')
       .mockImplementationOnce(() => [
-        EventFactory.fromProto({
+        InworldPacket.fromProto({
           ...audioEvent,
           packetId: {
             packetId: audioEvent.packetId.packetId,
@@ -559,7 +561,7 @@ describe('onMessage', () => {
     jest
       .spyOn(GrpcAudioPlayback.prototype, 'excludeCurrentInteractionPackets')
       .mockImplementationOnce(() => [
-        EventFactory.fromProto({
+        InworldPacket.fromProto({
           ...audioEvent,
           packetId: {
             ...textEvent.packetId,
@@ -589,7 +591,7 @@ describe('onMessage', () => {
     jest
       .spyOn(GrpcAudioPlayback.prototype, 'excludeCurrentInteractionPackets')
       .mockImplementationOnce(() => [
-        EventFactory.fromProto({
+        InworldPacket.fromProto({
           ...audioEvent,
           packetId: {
             ...textEvent.packetId,