+++ src/common/helpers.ts
@@ -1,0 +1,2 @@
+export const protoTimestamp = (date?: Date) =>
+  (date || new Date()).toISOString();